+++ src/agdb/collections/map/map_data_storage_index.rs
@@ -1,5 +1,5 @@
 use crate::db::db_error::DbError;
-use crate::storage::storage_index::StorageIndex;
+use crate::storage::StorageIndex;
 use crate::utilities::serialize::Serialize;
 use crate::utilities::serialize::SerializeStatic;
