+++ src/agdb/collections/map_storage.rs
@@ -4,8 +4,8 @@ use super::map::multi_map_impl::MultiMapImpl;
 use crate::collections::vec::VecValue;
 use crate::db::db_error::DbError;
 use crate::storage::file_storage::FileStorage;
-use crate::storage::storage_index::StorageIndex;
 use crate::storage::Storage;
+use crate::storage::StorageIndex;
 use crate::utilities::stable_hash::StableHash;
 use std::cell::RefCell;
 use std::collections::HashMap;
