+++ src/agdb/graph/graph_data_storage.rs
@@ -5,8 +5,8 @@ use super::graph_index::GraphIndex;
 use crate::collections::vec::DbVec;
 use crate::db::db_error::DbError;
 use crate::storage::file_storage::FileStorage;
-use crate::storage::storage_index::StorageIndex;
 use crate::storage::Storage;
+use crate::storage::StorageIndex;
 use std::cell::RefCell;
 use std::rc::Rc;
