+++ src/agdb/graph/graph_index.rs
@@ -78,30 +78,6 @@ impl VecValue for GraphIndex {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::cmp::Ordering;
-    use std::collections::hash_map::DefaultHasher;
-    use std::hash::Hash;
-    use std::hash::Hasher;
-
-    #[test]
-    fn derived_from_debug() {
-        let index = GraphIndex::default();
-
-        format!("{index:?}");
-    }
-
-    #[test]
-    fn derived_from_hash() {
-        let mut hasher = DefaultHasher::new();
-        GraphIndex { index: 1 }.hash(&mut hasher);
-        assert_ne!(hasher.finish(), 0);
-    }
-
-    #[test]
-    fn derived_from_ord() {
-        let index = GraphIndex::default();
-        assert_eq!(index.cmp(&index), Ordering::Equal);
-    }
 
     #[test]
     fn is_edge() {
@@ -115,28 +91,6 @@ mod tests {
         assert!(GraphIndex::from(1).is_node());
         assert!(!GraphIndex::default().is_node());
         assert!(!GraphIndex::from(-1).is_node());
-    }
-
-    #[test]
-    fn ordering() {
-        let mut indexes = vec![
-            GraphIndex::default(),
-            GraphIndex::from(100_i64),
-            GraphIndex::from(-1_i64),
-            GraphIndex::from(1_i64),
-        ];
-
-        indexes.sort();
-
-        assert_eq!(
-            indexes,
-            vec![
-                GraphIndex::from(-1_i64),
-                GraphIndex::default(),
-                GraphIndex::from(1_i64),
-                GraphIndex::from(100_i64),
-            ]
-        )
     }
 
     #[test]
