+++ src/agdb/query/query_condition/comparison.rs
@@ -19,20 +19,4 @@ mod tests {
     fn derived_from_debug() {
         format!("{:?}", Comparison::Equal(DbValue::Int(0)));
     }
-
-    #[test]
-    fn derived_from_clone() {
-        let left = Comparison::Equal(DbValue::Int(0));
-        let right = left.clone();
-
-        assert_eq!(left, right);
-    }
-
-    #[test]
-    fn derived_from_partial_eq() {
-        assert_eq!(
-            Comparison::Equal(DbValue::Int(0)),
-            Comparison::Equal(DbValue::Int(0))
-        );
-    }
 }
