+++ src/agdb/query/query_condition/direction.rs
@@ -14,17 +14,4 @@ mod tests {
     fn derived_from_debug() {
         format!("{:?}", Direction::From);
     }
-
-    #[test]
-    fn derived_from_clone() {
-        let left = Direction::From;
-        let right = left.clone();
-
-        assert_eq!(left, right);
-    }
-
-    #[test]
-    fn derived_from_partial_eq() {
-        assert_eq!(Direction::From, Direction::From);
-    }
 }
