+++ src/agdb/query/query_ids.rs
@@ -15,20 +15,4 @@ mod tests {
     fn derived_from_debug() {
         format!("{:?}", QueryIds::Ids(vec![QueryId::from(0)]));
     }
-
-    #[test]
-    fn derived_from_clone() {
-        let left = QueryIds::Ids(vec![QueryId::from(0)]);
-        let right = left.clone();
-
-        assert_eq!(left, right);
-    }
-
-    #[test]
-    fn derived_from_partial_eq() {
-        assert_eq!(
-            QueryIds::Ids(vec![QueryId::from(0)]),
-            QueryIds::Ids(vec![QueryId::from(0)])
-        );
-    }
 }
