+++ src/agdb/query/query_result.rs
@@ -5,13 +5,3 @@ pub struct QueryResult {
     pub result: i64,
     pub elements: Vec<DbElement>,
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn derived_from_debug() {
-        format!("{:?}", QueryResult::default());
-    }
-}
